refactor: restructure memo module with useMemoTab hook and utility functions
- Introduced useMemoTab hook for managing memo state and operations.
- Moved utility functions related to memo processing to memoUtils.ts.
- Created memoConfig.ts for storing constants and configurations.
- Defined types for memo items and hook return values in memoTypes.ts.
- Updated MemoTab component to utilize the new hook and utility functions.
- Improved code organization and readability by separating concerns.
